Attempt to utilize user data from query
