+++ src/components/App/App.tsx
@@ -7,31 +7,16 @@ import { Redirect, Route, Switch } from 'react-router-dom';
 import { NewConditionPage } from '../NewConditionPage/NewConditionPage';
 import { ErrorPage } from '../ErrorPage/ErrorPage';
 import { useQuery, gql } from '@apollo/client';
+import { useState } from 'react';
+import { Users } from '../../gql/graphql';
 
 export const App = () => {
-  const GET_USERS = gql`
-    query User {
-      user(id:1) {
-        id
-        name    
-      }
-    }
-  `
 
-  function DisplayUser() {
-    const { loading, error, data } = useQuery(GET_USERS);
-    if (loading) return console.log('Loading', loading)
-    if (error) return console.log('Error', error);
-    console.log(data)
-    return data
-  }
-
-  DisplayUser()
   return (
     <main>
       <NavBar />
       <Switch>
-        <Route exact path='/' render={() => <HomePage />} />
+        <Route exact path='/' render={() => <HomePage/>} />
         <Route exact path='/user-dashboard' render={() => <UserDashboard />} />
         <Route exact path='/conditions/:condition' render={({match}) => <ConditionPage key={match.params.condition}/>} />
         <Route exact path='/add-condition'render={() => <NewConditionPage />} />
